feat: split up validations and tests
