+++ __tests__/validations/component.test.tsx
@@ -1,0 +1,163 @@
+import test, { ExecutionContext } from 'ava';
+
+import { Component, Prefab } from '../../src/types';
+import validateComponents from '../../src/validations/component';
+import validatePrefabs from '../../src/validations/prefab';
+
+type Context = ExecutionContext<unknown>;
+
+test('Throw when one of the components is invalid', (t: Context): void => {
+  const components: { name: string }[] = [
+    {
+      name: 'HelloWorld',
+    },
+  ];
+
+  t.throws(() => validateComponents(components as Component[]));
+});
+
+test('Throw when two components have the same name', (t: Context): void => {
+  const components = [
+    {
+      name: 'HelloWorld',
+      type: 'ROW',
+      allowedTypes: ['COLUMN'],
+      orientation: 'VERTICAL',
+      jsx: '<div>jsx</div>',
+      styles: 'styles',
+    },
+    {
+      name: 'HelloWorld',
+      type: 'ROW',
+      allowedTypes: ['COLUMN'],
+      orientation: 'VERTICAL',
+      jsx: '<div>jsx</div>',
+      styles: 'styles',
+    },
+  ] as Component[];
+
+  t.throws(() => validateComponents(components));
+});
+
+test("Don't throw when all components are valid", (t: Context): void => {
+  const components = [
+    {
+      name: 'HelloWorld',
+      type: 'ROW',
+      allowedTypes: ['COLUMN'],
+      orientation: 'VERTICAL',
+      jsx: '<div>jsx</div>',
+      styles: 'styles',
+    },
+  ] as Component[];
+
+  t.notThrows(() => validateComponents(components));
+});
+
+test('Throw when one of the prefabs is invalid', (t: Context): void => {
+  const prefabs = [
+    {
+      name: 'Component Name',
+    },
+  ] as Prefab[];
+
+  t.throws(() => validatePrefabs(prefabs));
+});
+
+test("Don't throw when all prefabs are valid", (t: Context): void => {
+  const prefabs = [
+    {
+      name: 'Component Name',
+      icon: 'AccordionIcon',
+      category: 'Content',
+      structure: [],
+    },
+  ] as Prefab[];
+
+  t.notThrows(() => validatePrefabs(prefabs));
+});
+
+test('Throw when one of the prefabs options is invalid', (t: Context): void => {
+  const prefabs = ([
+    {
+      name: 'Component Name',
+      icon: 'TitleIcon',
+      category: 'CONTENT',
+      structure: [
+        {
+          name: 'something',
+          options: [
+            {
+              value: '',
+              label: 'something',
+              key: 'something',
+              type: 'TEXT',
+              invalid: ' ',
+            },
+          ],
+          descendants: [],
+        },
+      ],
+    },
+  ] as unknown) as Prefab[];
+
+  t.throws(() => validatePrefabs(prefabs));
+});
+
+test('Throw when the prefabs option type is not referring to one the correct types', (t: Context): void => {
+  const prefabs = [
+    {
+      name: 'Component Name',
+      icon: 'TitleIcon',
+      category: 'CONTENT',
+      structure: [
+        {
+          name: 'something',
+          options: [
+            {
+              value: '',
+              label: 'something',
+              key: 'something',
+              type: 'SOMETHING',
+            },
+          ],
+          descendants: [],
+        },
+      ],
+    },
+  ] as Prefab[];
+
+  t.throws(() => validatePrefabs(prefabs));
+});
+
+test('Throw when two options with the same key are being used', (t: Context): void => {
+  const prefabs = [
+    {
+      name: 'Component Name',
+      icon: 'TitleIcon',
+      category: 'CONTENT',
+      structure: [
+        {
+          name: 'something',
+          options: [
+            {
+              value: '',
+              label: 'Option 1',
+              key: 'sameKey',
+              type: 'TEXT',
+            },
+            {
+              value: '',
+              label: 'Option 2',
+              key: 'sameKEY',
+              type: 'TEXT',
+            },
+          ],
+          descendants: [],
+        },
+      ],
+    },
+  ] as Prefab[];
+
+  t.throws(() => validatePrefabs(prefabs));
+});
